T#1861/tags management

+++ service/src/main/java/skills/services/admin/SkillTagService.groovy
@@ -21,7 +21,9 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import skills.controller.exceptions.ErrorCode
 import skills.controller.exceptions.SkillException
+import skills.controller.request.model.SkillsTagDeleteRequest
 import skills.controller.request.model.SkillsTagRequest
+import skills.controller.result.model.SkillTagInfoRes
 import skills.services.LockingService
 import skills.services.RuleSetDefGraphService
 import skills.services.userActions.DashboardAction
@@ -82,25 +84,48 @@ class SkillTagService {
                 savedSkill = skillDefWithExtraRepo.save(skillDefinition)
             }
             log.debug("Saved [{}]", savedSkill)
+        } else {
+            if (skillsTagRequest.tagValue && skillsTagRequest.tagValue != skillDefinition.name) {
+                skillDefinition.name = skillsTagRequest.tagValue
+                SkillDefWithExtra savedSkill
+                DataIntegrityExceptionHandlers.tagDataIntegrityViolationExceptionHandler.handle(projectId) {
+                    savedSkill = skillDefWithExtraRepo.save(skillDefinition)
+                }
+                log.debug("Updated [{}]", savedSkill)
+            }
         }
 
         List<String> existingTaggedSkillIds = getSkillsForTag(projectId, skillsTagRequest.tagId)?.collect { it.skillId }
-        skillsTagRequest.skillIds.each { skillId ->
-            if (!existingTaggedSkillIds.find { it == skillId}) {
-                ruleSetDefGraphService.assignGraphRelationship(projectId, skillDefinition.skillId, SkillDef.ContainerType.Tag, skillId, SkillRelDef.RelationshipType.Tag)
-            }
 
+        if (skillsTagRequest.skillIds) {
+            skillsTagRequest.skillIds?.each { skillId ->
+                if (!existingTaggedSkillIds.find { it == skillId}) {
+                    ruleSetDefGraphService.assignGraphRelationship(projectId, skillDefinition.skillId, SkillDef.ContainerType.Tag, skillId, SkillRelDef.RelationshipType.Tag)
+                }
+
+                userActionsHistoryService.saveUserAction(new UserActionInfo(
+                        action: DashboardAction.Create,
+                        item: DashboardItem.Tag,
+                        actionAttributes: [
+                                tagId: skillsTagRequest.tagId,
+                                tagValue: skillsTagRequest.tagValue,
+                        ],
+                        itemId: skillId,
+                        projectId: projectId,
+                ))
+            }
+        } else {
             userActionsHistoryService.saveUserAction(new UserActionInfo(
                     action: DashboardAction.Create,
                     item: DashboardItem.Tag,
                     actionAttributes: [
-                            tagId: skillsTagRequest.tagId,
                             tagValue: skillsTagRequest.tagValue,
                     ],
-                    itemId: skillId,
+                    itemId: skillsTagRequest.tagId,
                     projectId: projectId,
             ))
         }
+
 
         log.debug("Added [{}] tag to skills [{}]", skillsTagRequest.tagValue, skillsTagRequest.skillIds)
     }
@@ -109,6 +134,35 @@ class SkillTagService {
     List<SkillTag> getTagsForProject(String projectId, Boolean includeDisabled=true) {
         skillDefRepo.getTagsForProject(projectId, includeDisabled.toString())
     }
+
+    @Transactional
+    SkillTagInfoRes getSingleTagInfo(String projectId, String tagId, Boolean includeDisabled=true) {
+
+        List<SkillDefPartial> skillsForTag = getSkillsForTag(projectId, tagId)
+
+        SkillDef tagSkillDef = skillDefRepo.findByProjectIdAndSkillIdAndType(projectId, tagId, SkillDef.ContainerType.Tag)
+        if (!tagSkillDef) {
+            throw new SkillException("Tag with id [${tagId}] does not exist.", projectId, null, ErrorCode.BadParam)
+        }
+
+        return new SkillTagInfoRes(
+                tagId: tagSkillDef.skillId,
+                tagValue: tagSkillDef.name,
+                skills: skillsForTag?.sort { it.relationshipCreated }?.reversed().collect {
+                    new SkillTagInfoRes.SkillInfo(
+                            skillId: it.skillId,
+                            skillName: it.name,
+                            subjectName: it.subjectName,
+                            subjectId: it.subjectSkillId,
+                            groupName: it.groupName,
+                            groupId: it.groupId,
+                            taggedOn: it.relationshipCreated,
+                    )
+                }
+        )
+
+    }
+
 
     @Transactional
     List<SkillTag> getTagsForSubject(String projectId, String subjectId, Boolean includeDisabled=true) {
@@ -136,7 +190,7 @@ class SkillTagService {
     }
 
     @Transactional
-    void deleteTagForSkills(String projectId, SkillsTagRequest skillsTagRequest) {
+    void deleteTagForSkills(String projectId, SkillsTagDeleteRequest skillsTagRequest) {
         String tagId = skillsTagRequest.tagId
         SkillDefWithExtra tag = skillDefWithExtraRepo.findByProjectIdAndSkillIdIgnoreCaseAndType(projectId, tagId, SkillDef.ContainerType.Tag)
 
@@ -144,24 +198,45 @@ class SkillTagService {
             throw new SkillException("Tag [${tagId}] does not exist", projectId, tagId, ErrorCode.TagNotFound)
         }
 
-        skillsTagRequest.skillIds.each {skillId ->
+        boolean removeTagFully = skillsTagRequest.removeTagFully == true
+        List<String> skillIds = (removeTagFully) ? getSkillsForTag(projectId, tagId)?.skillId : skillsTagRequest.skillIds
+
+        skillIds.each {skillId ->
             ruleSetDefGraphService.removeGraphRelationship(projectId, tagId, SkillDef.ContainerType.Tag,
                     projectId, skillId, SkillRelDef.RelationshipType.Tag, false)
+        }
 
+        if (removeTagFully) {
             userActionsHistoryService.saveUserAction(new UserActionInfo(
                     action: DashboardAction.Delete,
                     item: DashboardItem.Tag,
                     actionAttributes: [
-                            tagId: tag.skillId,
                             tagValue: tag.name,
+                            removedFromSkills: skillIds,
                     ],
-                    itemId: skillId,
+                    itemId: tagId,
                     projectId: projectId,
             ))
+        } else {
+            skillIds.each {skillId ->
+                userActionsHistoryService.saveUserAction(new UserActionInfo(
+                        action: DashboardAction.Delete,
+                        item: DashboardItem.Tag,
+                        actionAttributes: [
+                                tagId: tag.skillId,
+                                tagValue: tag.name,
+                        ],
+                        itemId: skillId,
+                        projectId: projectId,
+                ))
+            }
         }
 
         Integer skillsWithTag = skillRelDefRepo.getSkillWithTagCount(tagId)
-        if (skillsWithTag <= 0) {
+        if (removeTagFully) {
+            assert skillsWithTag <= 0
+        }
+        if (!skillsTagRequest.retainTag && skillsWithTag <= 0) {
             log.info("Tag [${tagId}] removed from all skills, removing tag definition")
             skillDefWithExtraRepo.delete(tag)
         }
